% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{all_ident}
\alias{all_ident}
\title{Create parset object with identity transformed parameters}
\usage{
all_ident(...)
}
\arguments{
\item{...}{named values; see example}
}
\description{
Create parset object with identity transformed parameters
}
