% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{all_log}
\alias{all_log}
\title{Create parset object with all log-transformed parameters}
\usage{
all_log(...)
}
\arguments{
\item{...}{named values; see example}
}
\description{
Create parset object with all log-transformed parameters
}
\examples{

all_log(CL = 1.1, V2 = 20, Q = 3.5, V3 = 201)

}
