% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{as_par}
\alias{as_par}
\alias{as_par_logit}
\alias{as_par_fixed}
\alias{as_par_ident}
\title{Create parameter transformation objects}
\usage{
as_par(..., fun = log_par)

as_par_logit(...)

as_par_fixed(...)

as_par_ident(...)
}
\arguments{
\item{...}{passed to [as_par]}

\item{fun}{the transformation-specific constructor function}

\item{a}{single named numeric value (see examples)}

\item{arguments}{passed to [as_par]}
}
\description{
Create parameter transformation objects
}
\examples{

as_par(CL = 2)

as_par_fixed(g = 9.8)

}
