% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{coef,parset-method}
\alias{coef,parset-method}
\title{Get estimated (non-fixed) values from a parset object.}
\usage{
\S4method{coef}{parset}(object, all = FALSE, ...)
}
\arguments{
\item{object}{the parset object}

\item{all}{return non-fixed values too}

\item{...}{not used}
}
\description{
Get estimated (non-fixed) values from a parset object.
}
\seealso{
\code{\link{coeff}}
}
