% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coeff}
\alias{coeff}
\title{Get estimated and fixed values from a parset object.}
\usage{
coeff(x)
}
\arguments{
\item{x}{a parset object}
}
\description{
Get estimated and fixed values from a parset object.
}

