% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{construct_par}
\alias{construct_par}
\alias{log_par}
\alias{logit_par}
\alias{fixed_par}
\alias{ident_par}
\alias{new_par}
\title{Parameter constructor functions}
\usage{
log_par(name, value, fixed = FALSE)

logit_par(name, value, fixed = FALSE)

fixed_par(name, value, fixed = TRUE)

ident_par(name, value, fixed = FALSE)

new_par(name, value, tr, un, fixed = FALSE, trans = FALSE)
}
\arguments{
\item{name}{character name of the parameter}

\item{value}{the initial parameter value}

\item{fixed}{if `TRUE`, the parameter will be fixed}
}
\description{
Parameter constructor functions
}
