% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_nl}
\alias{fit_nl}
\title{Fit a model to data using functions from nloptr}
\usage{
fit_nl(theta, data, mod, optimizer = c("newuoa", "neldermead"),
  pred_name = "CP", ofv = els, logdv = FALSE, sigma = ~sigma,
  pred_initial = FALSE, cov_step = FALSE, ...)
}
\arguments{
\item{theta}{a parset object}

\item{data}{a data set in data.frame format}

\item{mod}{a mrgsolve model object}

\item{optimizer}{used to search the parameter space}

\item{pred_name}{name in simulated output representing `PRED`}

\item{ofv}{objective function function}

\item{logdv}{logical; if `TRUE`, then the dependent variable is 
log-transformed during `ofv` calculation}

\item{sigma}{a one-sided formula containing an expression for the residual
error variance}

\item{pred_initial}{if `TRUE`, a column is added to the output called
`INITIAL` which is a model prediction at the initial parameter estimates}

\item{cov_step}{if `TRUE`, then standard errors are derived using 
the Hessian}

\item{...}{passed to `ofv`}
}
\description{
Fit a model to data using functions from nloptr
}
