% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{fix_par}
\alias{fix_par}
\title{Fix some parameters in a set.}
\usage{
fix_par(x, ...)
}
\arguments{
\item{x}{parset object}

\item{...}{unquoted names of paramters to fix}
}
\description{
Fix some parameters in a set.
}

