% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{get_trans}
\alias{get_trans}
\alias{get_untrans}
\alias{get_initials}
\alias{initials}
\alias{get_pars}
\title{Get values from a parset object}
\usage{
get_trans(x)

get_untrans(x)

get_initials(x)

initials(x)

get_pars(x)
}
\arguments{
\item{x}{a parset object}
}
\description{
Get values from a parset object
}
\examples{
x <- quick_par(CL = log(1), V2 = log(20), F1 = logit(0.77))

get_trans(x)

get_untrans(x)

}
