% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{graft}
\alias{graft}
\alias{graft_par}
\alias{graft_in}
\title{Graft new values into an existing parset object}
\usage{
graft(x, y)

graft_par(x, y)

graft_in(x, y)

graft_in(x, y)
}
\arguments{
\item{x}{a parset object}

\item{y}{numeric vector of new values}
}
\description{
Graft new values into an existing parset object
}
