% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{graft__}
\alias{graft__}
\title{Graft new estimates into a \code{parset} object.}
\usage{
graft__(x, y, untrans = TRUE, ...)
}
\arguments{
\item{x}{a parset object}

\item{y}{named list or vector of estimates to graft into the parlist object}

\item{untrans}{logical indicating whether to untransform the object before returning}

\item{...}{not used}
}
\description{
Graft new estimates into a \code{parset} object.
}
