% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{ident_par}
\alias{ident_par}
\title{Create a list of untransformed parameters.}
\usage{
ident_par(name, value, ...)
}
\arguments{
\item{name}{the parameter name}

\item{value}{the parameter value}

\item{...}{passed to constructor}
}
\description{
Create a list of untransformed parameters.
}

