% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimhelp.R
\name{log_par}
\alias{log_par}
\title{Create a list of log-transformed parameters.}
\usage{
log_par(name, value, ...)
}
\arguments{
\item{name}{the parameter name}

\item{value}{the parameter value}

\item{...}{passed to constructor}
}
\description{
Create a list of log-transformed parameters.
}
