% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{nl_optr}
\alias{nl_optr}
\title{Fit a model to data}
\usage{
nl_optr(theta, data, optimizer = c("newuoa", "neldermead", "optim"),
  pred_name = "CP", ofv = els, logdv = FALSE, pred_initial = FALSE,
  cov_step = FALSE, ...)
}
\arguments{
\item{theta}{a parset object}

\item{data}{a data set in data.frame format}

\item{optimizer}{used to search the parameter space}

\item{pred_name}{name in simulated output representing `PRED`}

\item{ofv}{objective function function}

\item{logdv}{logical; if `TRUE`, then the dependent variable is 
log-transformed during `ofv` calculation}

\item{cov_step}{if `TRUE`, then standard errors are derived using 
the Hessian}

\item{...}{not currently used}
}
\description{
Fit a model to data
}
