% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofv.R
\name{ofv}
\alias{ofv}
\alias{els}
\alias{ml}
\alias{ols}
\alias{wls}
\title{Objective functions}
\usage{
els(dv, pred, sigma, logdv = FALSE, ...)

ml(dv, pred, sigma, logdv = FALSE, ...)

ols(dv, pred, ...)

wls(dv, pred, ...)
}
\arguments{
\item{dv}{observed data vector}

\item{pred}{model-predicted data vector}

\item{sigma}{residual error variance; may be length 1 or length of the
\code{dv} and \code{pred} vectors}

\item{logdv}{if \code{TRUE}, then observations and predictions will be
log-transformed prior to calculating the objective function value}

\item{...}{not used}
}
\description{
These functions take in observed and predicted data as well as other
data items and return a single objective function value quantifying
how similar the predicted data are to the observed data.
}
\details{
\code{els}: extended least squares; \code{ml}: maximum likelihood (normal);
\code{ols}: ordinary least squares; \code{wls}: weighted least squares
}
