% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{par}
\alias{par}
\alias{update,par-method}
\title{Par object class.}
\usage{
\S4method{update}{par}(object, value, trans = TRUE, ...)
}
\arguments{
\item{object}{par object}

\item{value}{the initial, untransformed value of the parameter}

\item{trans}{the parameter transformation}

\item{...}{passed along}

\item{x}{par object}
}
\description{
Par object class.
}
