% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/optimhelp.R
\docType{methods}
\name{is.parset}
\alias{is.parset}
\alias{is.par}
\alias{parset}
\alias{as.numeric,parset-method}
\alias{as.list,parset-method}
\alias{names,parset-method}
\alias{show,parset-method}
\alias{$,parset-method}
\alias{parset}
\title{Parset object class.}
\usage{
is.parset(x)

is.par(x)

\S4method{as.numeric}{parset}(x, ...)

\S4method{as.list}{parset}(x, ...)

\S4method{names}{parset}(x)

\S4method{show}{parset}(object)

\S4method{$}{parset}(x, name)

parset(...)
}
\arguments{
\item{x}{par object}

\item{...}{par objects}

\item{object}{par object}

\item{name}{parameter name}
}
\description{
Parset object class.

Create a new pars object.
}
