% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{quick_par}
\alias{quick_par}
\title{Quick and easy syntax for constructing parlist}
\usage{
quick_par(...)
}
\arguments{
\item{...}{named value, with an optional transformation of the value; 
valid transformations include `log`, `logit`, `fixed`, `ident`; see
examples}
}
\description{
Quick and easy syntax for constructing parlist
}
\examples{

quick_par(CL = log(1), KA = fixed(1.0), F1 = logit(0.8))

}
