% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{tform}
\alias{tform}
\alias{tform,par-method}
\alias{tform,logpar-method}
\alias{tform,logitpar-method}
\alias{tform,parset-method}
\title{Get the name of the transformation function.}
\usage{
tform(x, ...)

\S4method{tform}{par}(x, ...)

\S4method{tform}{logpar}(x, ...)

\S4method{tform}{logitpar}(x, ...)

\S4method{tform}{parset}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{not used}
}
\description{
Get the name of the transformation function.
}
