% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{trans}
\alias{trans}
\alias{untrans}
\title{Transform and un-transform a parset object}
\usage{
trans(x)

untrans(x)
}
\arguments{
\item{x}{a parset object}
}
\description{
Transform and un-transform a parset object
}
\examples{
x <- quick_par(CL = log(2))

x

trans(x)

untrans(trans(x))

}
