% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{untrans}
\alias{untrans}
\alias{untrans,par-method}
\alias{untrans,logpar-method}
\alias{untrans,logitpar-method}
\alias{untrans,parset-method}
\title{Untransform parameters in a parset object.}
\usage{
untrans(x, ...)

\S4method{untrans}{par}(x, ...)

\S4method{untrans}{logpar}(x, ...)

\S4method{untrans}{logitpar}(x, ...)

\S4method{untrans}{parset}(x, ...)
}
\arguments{
\item{x}{a par or parset object}

\item{...}{not used}
}
\description{
Untransform parameters in a parset object.
}
