% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{apply_trans}
\alias{apply_trans}
\alias{apply_untrans}
\title{Apply transformations to new values}
\usage{
apply_trans(x, y)

apply_untrans(x, y)
}
\arguments{
\item{x}{a parset object}

\item{y}{a vector of numeric values for transformation or 
untransformation}
}
\description{
Apply transformations to new values
}
