% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{as_par}
\alias{as_par}
\alias{as_par_logit}
\alias{as_par_fixed}
\alias{all_log}
\alias{all_ident}
\alias{as_par_ident}
\alias{log_par}
\alias{logit_par}
\alias{fixed_par}
\alias{ident_par}
\alias{new_par}
\title{Create parameter transformation objects}
\usage{
as_par(..., fun = log_par)

as_par_logit(...)

as_par_fixed(...)

all_log(...)

all_ident(...)

as_par_ident(...)

log_par(name, value, fixed = FALSE)

logit_par(name, value, fixed = FALSE)

fixed_par(name, value, fixed = TRUE)

ident_par(name, value, fixed = FALSE)

new_par(name, value, tr, un, fixed = FALSE, trans = FALSE)
}
\arguments{
\item{fun}{the transformation-specific constructor function}

\item{a}{single named numeric value (see examples)}
}
\description{
Create parameter transformation objects
}
\examples{

as_par(CL = 2)

as_par_fixed(g = 9.8)

}
