% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlist.R
\name{parset_add}
\alias{parset_add}
\title{Add a par object to a parset object}
\usage{
parset_add(x, nw)
}
\arguments{
\item{a}{parset object}

\item{a}{par object}
}
\description{
Add a par object to a parset object
}
