% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{trans}
\alias{trans}
\alias{trans,par-method}
\alias{trans,logpar-method}
\alias{trans,logitpar-method}
\alias{trans,parset-method}
\title{Transform transformed parameters in a parset object.}
\usage{
trans(x, ...)

\S4method{trans}{par}(x, ...)

\S4method{trans}{logpar}(x, ...)

\S4method{trans}{logitpar}(x, ...)

\S4method{trans}{parset}(x, ...)
}
\arguments{
\item{x}{a par or parset object}

\item{...}{not used}
}
\description{
Transform transformed parameters in a parset object.
}
